add some layout for project details setup w/ tabs
