+++ client/src/components/grants/Rounds.tsx
@@ -2,18 +2,8 @@ import { useSelector } from "react-redux";
 import { useParams } from "react-router-dom";
 import { useNetwork } from "wagmi";
 import { RootState } from "../../reducers";
-import { FormInputs, Metadata, Project } from "../../types";
-import ProjectDetailsHeader from "../ProjectDetailsHeader";
 
-export default function Rounds({
-  project,
-  bannerImg,
-  logoImg,
-}: {
-  project?: Metadata | FormInputs | Project;
-  bannerImg: string | Blob;
-  logoImg: string | Blob;
-}) {
+export default function Rounds() {
   const params = useParams();
   const { chain } = useNetwork();
   const props = useSelector((state: RootState) => {
@@ -30,12 +20,8 @@ export default function Rounds({
 
   return (
     <div className="w-full mb-40">
-      <ProjectDetailsHeader
-        title={project?.title}
-        bannerImg={bannerImg}
-        logoImg={logoImg}
-      />
-      {/* add new tab layout here */}
+      {/* list the application details here for each round */}
+      <span>Hello World!!!</span>
     </div>
   );
 }
