+++ client/src/components/grants/Stats.tsx
@@ -1,0 +1,7 @@
+export default function RoundStats() {
+  return (
+    <div>
+      <span>Hello World!!!</span>
+    </div>
+  );
+}
